requirements.txt
mysql-connector-python>=8.0.0
scikit-learn>=1.0.0
matplotlib>=3.5.0
reportlab>=3.6.0
numpy>=1.21.0